+++ changelog.md
@@ -1,3 +1,9 @@
+1.0.1 (2016-08-26)
+------------------
+
+- Fix UniqueObject constraint comparison
+- Remove short array syntax (PHP 5.4) to be compatible with PHP 5.3
+
 1.0.0 (2016-08-23)
 ------------------